pyproject.toml
[tool.poetry]
name = "test-ocr"
version = "0.1.0"
description = ""
authors = ["Julius Dehner <julius.dehner@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyQt5 = "^5.15.4"
pyscreenshot = "^3.0"
Pillow = "^8.3.1"
numpy = "^1.21.1"
pytesseract = "^0.3.8"
pyperclip = "^1.8.2"
pykakasi = "^2.2.1"
deep-translator = "^1.5.0"
gTTS = "^2.2.3"
pydub = "^0.25.1"
notify-py = "^0.3.3"
typer = "^0.3.2"
pynput = "^1.7.3"
xlib = "^0.21"
playsound = "^1.3.0"

[tool.poetry.dev-dependencies]
PyQt5-stubs = "^5.15.2"
mypy = "^0.910"
flake8 = "^3.9.2"
types-Pillow = "^8.3.1"
pyinstaller = "^4.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
